# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/registries": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve configuration settings for all supported registry formats",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get all registry settings",
                "responses": {
                    "200": {
                        "description": "Registry settings retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lgulliver_lodestone_pkg_types.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_lgulliver_lodestone_pkg_types.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Admin privileges required",
                        "schema": {
                            "$ref": "#/definitions/github_com_lgulliver_lodestone_pkg_types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve registry settings",
                        "schema": {
                            "$ref": "#/definitions/github_com_lgulliver_lodestone_pkg_types.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/registries/{registry}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve configuration settings for a specific registry format",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get specific registry setting",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Registry name (e.g., npm, nuget, maven)",
                        "name": "registry",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registry setting retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lgulliver_lodestone_pkg_types.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_lgulliver_lodestone_pkg_types.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Admin privileges required",
                        "schema": {
                            "$ref": "#/definitions/github_com_lgulliver_lodestone_pkg_types.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Registry not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_lgulliver_lodestone_pkg_types.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_lgulliver_lodestone_pkg_types.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}